@@ -3,35 +3,33 @@ description: This app is a clone of google translate using their api keys
 repository: https://github.com/camilo1498/flutter_translate_app
 homepage: https://github.com/camilo1498
 publish_to: 'none'
-version: 1.0.25
+version: 1.0.0+26
 
 environment:
-  sdk: ">=2.16.1 <3.0.0"
+  sdk: '>=3.3.4 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
-  cupertino_icons: ^1.0.2
-  provider: ^6.0.2
-  grouped_list: ^4.2.0
-  intl: ^0.17.0
-  flutter_screenutil: ^5.3.1
+  provider: ^6.1.2
+  grouped_list: ^5.1.2
+  intl: ^0.19.0
+  flutter_screenutil: ^5.9.0
   flutter_keyboard_visibility: ^5.2.0
-  dio: ^4.0.5
-  sqflite: ^2.0.2
-  fluttertoast: ^8.0.9
-  flutter_slidable: ^1.2.0
-  flutter_tts: ^3.3.3
+  dio: ^5.4.3+1
+  sqflite: ^2.3.3+1
+  fluttertoast: ^8.2.5
+  flutter_slidable: ^3.1.0
+  flutter_tts: ^4.0.2
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
 
-  flutter_lints: ^1.0.0
+  flutter_lints: ^3.0.0
 
 
 flutter:
-
   uses-material-design: true
