@@ -10,15 +10,10 @@ import 'package:provider/provider.dart';
 
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
-  Paint.enableDithering = true;
 
   /// force portrait mode
   SystemChrome.setPreferredOrientations(
       [DeviceOrientation.portraitUp, DeviceOrientation.portraitDown]);
-
-  /// hide bottom navigation bar in android
-  // SystemChrome.setEnabledSystemUIMode(SystemUiMode.manual,
-  //     overlays: [SystemUiOverlay.top]);
 
   /// force transparent top bar
   SystemChrome.setSystemUIOverlayStyle(const SystemUiOverlayStyle(
@@ -49,15 +44,10 @@ class _MyAppState extends State<MyApp> {
       designSize: const Size(1080, 1920),
       minTextAdapt: true,
       splitScreenMode: true,
-      builder: () => MaterialApp(
+      builder: (context, child) => MaterialApp(
           title: 'Flutter translate app',
           debugShowCheckedModeBanner: false,
-          theme: ThemeData(
-              primarySwatch: Colors.red),
-          builder: (context, widget){
-            ScreenUtil.setContext(context);
-            return widget!;
-          },
+          theme: ThemeData(primarySwatch: Colors.red),
           home: NotificationListener<OverscrollIndicatorNotification>(
             onNotification: (overscroll) {
               overscroll.disallowIndicator();
