@@ -1,29 +1,15 @@
-// ignore_for_file: constant_identifier_names
-
-/*
-Name: Akshath Jain
-Date: 3/18/2019 - 4/2/2020
-Purpose: Defines the sliding_up_panel widget
-Copyright: © 2020, Akshath Jain. All rights reserved.
-Licensing: More information can be found here: https://github.com/akshathjain/sliding_up_panel/blob/master/LICENSE
-
-added modification to support down scroll controller when Slide direction is "DOWN" in line
-modified by: Camilo Velandia
-github: https://github.com/camilo1498
-*/
+import 'dart:math';
 
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
-import 'dart:math';
-
 import 'package:flutter/physics.dart';
 
 enum SlideDirection {
-  UP,
-  DOWN,
+  up,
+  down,
 }
 
-enum PanelState { OPEN, CLOSED }
+enum PanelState { open, closed }
 
 class SlidingUpPanel extends StatefulWidget {
   /// The Widget that slides into view. When the
@@ -198,8 +184,8 @@ class SlidingUpPanel extends StatefulWidget {
       this.parallaxEnabled = false,
       this.parallaxOffset = 0.1,
       this.isDraggable = true,
-      this.slideDirection = SlideDirection.UP,
-      this.defaultPanelState = PanelState.CLOSED,
+      this.slideDirection = SlideDirection.up,
+      this.defaultPanelState = PanelState.closed,
       this.header,
       this.footer})
       : assert(panel != null || panelBuilder != null),
@@ -228,7 +214,7 @@ class _SlidingUpPanelState extends State<SlidingUpPanel>
     _ac = AnimationController(
         vsync: this,
         duration: const Duration(milliseconds: 300),
-        value: widget.defaultPanelState == PanelState.CLOSED
+        value: widget.defaultPanelState == PanelState.closed
             ? 0.0
             : 1.0 //set the default panel state (i.e. set initial value of _ac)
         )
@@ -257,7 +243,7 @@ class _SlidingUpPanelState extends State<SlidingUpPanel>
   @override
   Widget build(BuildContext context) {
     return Stack(
-      alignment: widget.slideDirection == SlideDirection.UP
+      alignment: widget.slideDirection == SlideDirection.up
           ? Alignment.bottomCenter
           : Alignment.topCenter,
       children: <Widget>[
@@ -286,7 +272,7 @@ class _SlidingUpPanelState extends State<SlidingUpPanel>
                 onVerticalDragEnd: widget.backdropTapClosesPanel
                     ? (DragEndDetails dets) {
                         // only trigger a close if the drag is towards panel close position
-                        if ((widget.slideDirection == SlideDirection.UP
+                        if ((widget.slideDirection == SlideDirection.up
                                     ? 1
                                     : -1) *
                                 dets.velocity.pixelsPerSecond.dy >
@@ -340,10 +326,10 @@ class _SlidingUpPanelState extends State<SlidingUpPanel>
                     children: <Widget>[
                       //open panel
                       Positioned(
-                          top: widget.slideDirection == SlideDirection.UP
+                          top: widget.slideDirection == SlideDirection.up
                               ? 0.0
                               : null,
-                          bottom: widget.slideDirection == SlideDirection.DOWN
+                          bottom: widget.slideDirection == SlideDirection.down
                               ? 0.0
                               : null,
                           width: MediaQuery.of(context).size.width -
@@ -361,11 +347,11 @@ class _SlidingUpPanelState extends State<SlidingUpPanel>
                       // header
                       widget.header != null
                           ? Positioned(
-                              top: widget.slideDirection == SlideDirection.UP
+                              top: widget.slideDirection == SlideDirection.up
                                   ? 0.0
                                   : null,
                               bottom:
-                                  widget.slideDirection == SlideDirection.DOWN
+                                  widget.slideDirection == SlideDirection.down
                                       ? 0.0
                                       : null,
                               child: widget.header ?? const SizedBox(),
@@ -375,11 +361,11 @@ class _SlidingUpPanelState extends State<SlidingUpPanel>
                       // footer
                       widget.footer != null
                           ? Positioned(
-                              top: widget.slideDirection == SlideDirection.UP
+                              top: widget.slideDirection == SlideDirection.up
                                   ? null
                                   : 0.0,
                               bottom:
-                                  widget.slideDirection == SlideDirection.DOWN
+                                  widget.slideDirection == SlideDirection.down
                                       ? null
                                       : 0.0,
                               child: widget.footer ?? const SizedBox())
@@ -387,10 +373,10 @@ class _SlidingUpPanelState extends State<SlidingUpPanel>
 
                       // collapsed panel
                       Positioned(
-                        top: widget.slideDirection == SlideDirection.UP
+                        top: widget.slideDirection == SlideDirection.up
                             ? 0.0
                             : null,
-                        bottom: widget.slideDirection == SlideDirection.DOWN
+                        bottom: widget.slideDirection == SlideDirection.down
                             ? 0.0
                             : null,
                         width: MediaQuery.of(context).size.width -
@@ -431,7 +417,7 @@ class _SlidingUpPanelState extends State<SlidingUpPanel>
   }
 
   double _getParallax() {
-    if (widget.slideDirection == SlideDirection.UP) {
+    if (widget.slideDirection == SlideDirection.up) {
       return -_ac.value *
           (widget.maxHeight - widget.minHeight) *
           widget.parallaxOffset;
@@ -476,7 +462,7 @@ class _SlidingUpPanelState extends State<SlidingUpPanel>
   void _onGestureSlide(double dy) {
     // only slide the panel if scrolling is not enabled
     if (!_scrollingEnabled) {
-      if (widget.slideDirection == SlideDirection.UP) {
+      if (widget.slideDirection == SlideDirection.up) {
         _ac.value -= dy / (widget.maxHeight - widget.minHeight);
       } else {
         _ac.value += dy / (widget.maxHeight - widget.minHeight);
@@ -489,7 +475,7 @@ class _SlidingUpPanelState extends State<SlidingUpPanel>
     if (_isPanelOpen && _sc.hasClients && _sc.offset <= 0) {
       setState(() {
         /// if SlideDirection.UP scroll direction will be down
-        if (widget.slideDirection == SlideDirection.UP) {
+        if (widget.slideDirection == SlideDirection.up) {
           if (dy < 0) {
             _scrollingEnabled = true;
           } else {
@@ -526,7 +512,7 @@ class _SlidingUpPanelState extends State<SlidingUpPanel>
         -v.pixelsPerSecond.dy / (widget.maxHeight - widget.minHeight);
 
     // reverse visual velocity to account for slide direction
-    if (widget.slideDirection == SlideDirection.DOWN) {
+    if (widget.slideDirection == SlideDirection.down) {
       visualVelocity = -visualVelocity;
     }
 
